requirements.txt
fastapi
uvicorn
python-dotenv
requests
httpx
motor
dnspython
authlib
itsdangerous
jinja2
python-multipart
pydantic
groq
duckduckgo-search
google-generativeai
pypdf2
python-docx
pillow
pdf2docx
opencv-python-headless
pinecone
numpy
passlib[bcrypt]
email-validator
passlib==1.7.4
bcrypt==4.0.1
gradio_client>=1.3.0
youtube-transcript-api